Reorganize module hierarchy and public uses
