+++ src/manual_types.rs
@@ -1,8 +1,4 @@
 use tl::dynamic::TLObject;
-//use tl::Bare;
 
-//pub type BareVec<T> = Bare<Vec<T>>;
-//pub type Int128 = (i64, i64);
-//pub type Int256 = (Int128, Int128);
 pub type Object = Box<TLObject>;
 //pub use tl::dynamic::LengthAndObject;
